fix bug in PeSTO and fix github unit tests

+++ src/pesto.zig
@@ -10,7 +10,7 @@ inline fn flip(square: u7) u6 {
     return @truncate(u6, square) ^ 56;
 }
 
-const MIDGAME_PAWN_TABLE = [64]i16{
+const MIDGAME_PAWN_TABLE = [64]i32{
     0,   0,   0,   0,   0,   0,   0,  0,
     98,  134, 61,  95,  68,  126, 34, -11,
     -6,  7,   26,  31,  65,  56,  25, -20,
@@ -21,7 +21,7 @@ const MIDGAME_PAWN_TABLE = [64]i16{
     0,   0,   0,   0,   0,   0,   0,  0,
 };
 
-const ENDGAME_PAWN_TABLE = [64]i16{
+const ENDGAME_PAWN_TABLE = [64]i32{
     0,   0,   0,   0,   0,   0,   0,   0,
     178, 173, 158, 134, 147, 132, 165, 187,
     94,  100, 85,  67,  56,  53,  82,  84,
@@ -32,7 +32,7 @@ const ENDGAME_PAWN_TABLE = [64]i16{
     0,   0,   0,   0,   0,   0,   0,   0,
 };
 
-const MIDGAME_KNIGHT_TABLE = [64]i16{
+const MIDGAME_KNIGHT_TABLE = [64]i32{
     -167, -89, -34, -49, 61,  -97, -15, -107,
     -73,  -41, 72,  36,  23,  62,  7,   -17,
     -47,  60,  37,  65,  84,  129, 73,  44,
@@ -43,7 +43,7 @@ const MIDGAME_KNIGHT_TABLE = [64]i16{
     -105, -21, -58, -33, -17, -28, -19, -23,
 };
 
-const ENDGAME_KNIGHT_TABLE = [64]i16{
+const ENDGAME_KNIGHT_TABLE = [64]i32{
     -58, -38, -13, -28, -31, -27, -63, -99,
     -25, -8,  -25, -2,  -9,  -25, -24, -52,
     -24, -20, 10,  9,   -1,  -9,  -19, -41,
@@ -54,7 +54,7 @@ const ENDGAME_KNIGHT_TABLE = [64]i16{
     -29, -51, -23, -15, -22, -18, -50, -64,
 };
 
-const MIDGAME_BISHOP_TABLE = [64]i16{
+const MIDGAME_BISHOP_TABLE = [64]i32{
     -29, 4,  -82, -37, -25, -42, 7,   -8,
     -26, 16, -18, -13, 30,  59,  18,  -47,
     -16, 37, 43,  40,  35,  50,  37,  -2,
@@ -65,7 +65,7 @@ const MIDGAME_BISHOP_TABLE = [64]i16{
     -33, -3, -14, -21, -13, -12, -39, -21,
 };
 
-const ENDGAME_BISHOP_TABLE = [64]i16{
+const ENDGAME_BISHOP_TABLE = [64]i32{
     -14, -21, -11, -8,  -7, -9,  -17, -24,
     -8,  -4,  7,   -12, -3, -13, -4,  -14,
     2,   -8,  0,   -1,  -2, 6,   0,   4,
@@ -76,7 +76,7 @@ const ENDGAME_BISHOP_TABLE = [64]i16{
     -23, -9,  -23, -5,  -9, -16, -5,  -17,
 };
 
-const MIDGAME_ROOK_TABLE = [64]i16{
+const MIDGAME_ROOK_TABLE = [64]i32{
     32,  42,  32,  51,  63, 9,  31,  43,
     27,  32,  58,  62,  80, 67, 26,  44,
     -5,  19,  26,  36,  17, 45, 61,  16,
@@ -87,7 +87,7 @@ const MIDGAME_ROOK_TABLE = [64]i16{
     -19, -13, 1,   17,  16, 7,  -37, -26,
 };
 
-const ENDGAME_ROOK_TABLE = [64]i16{
+const ENDGAME_ROOK_TABLE = [64]i32{
     13, 10, 18, 15, 12, 12,  8,   5,
     11, 13, 13, 11, -3, 3,   8,   3,
     7,  7,  7,  5,  4,  -3,  -5,  -3,
@@ -98,7 +98,7 @@ const ENDGAME_ROOK_TABLE = [64]i16{
     -9, 2,  3,  -1, -5, -13, 4,   -20,
 };
 
-const MIDGAME_QUEEN_TABLE = [64]i16{
+const MIDGAME_QUEEN_TABLE = [64]i32{
     -28, 0,   29,  12,  59,  44,  43,  45,
     -24, -39, -5,  1,   -16, 57,  28,  54,
     -13, -17, 7,   8,   29,  56,  47,  57,
@@ -109,7 +109,7 @@ const MIDGAME_QUEEN_TABLE = [64]i16{
     -1,  -18, -9,  10,  -15, -25, -31, -50,
 };
 
-const ENDGAME_QUEEN_TABLE = [64]i16{
+const ENDGAME_QUEEN_TABLE = [64]i32{
     -9,  22,  22,  27,  27,  19,  10,  20,
     -17, 20,  32,  41,  58,  25,  30,  0,
     -20, 6,   9,   49,  47,  35,  19,  9,
@@ -120,7 +120,7 @@ const ENDGAME_QUEEN_TABLE = [64]i16{
     -33, -28, -22, -43, -5,  -32, -20, -41,
 };
 
-const MIDGAME_KING_TABLE = [64]i16{
+const MIDGAME_KING_TABLE = [64]i32{
     -65, 23,  16,  -15, -56, -34, 2,   13,
     29,  -1,  -20, -7,  -8,  -4,  -38, -29,
     -9,  24,  2,   -16, -20, 6,   22,  -22,
@@ -131,12 +131,12 @@ const MIDGAME_KING_TABLE = [64]i16{
     -15, 36,  12,  -54, 8,   -28, 24,  14,
 };
 
-const ENDGAME_KING_TABLE = [64]i16{ -74, -35, -18, -18, -11, 15, 4, -17, -12, 17, 14, 17, 17, 38, 23, 11, 10, 17, 23, 15, 20, 45, 44, 13, -8, 22, 24, 27, 26, 33, 26, 3, -18, -4, 21, 24, 27, 23, 9, -11, -19, -3, 11, 21, 23, 16, 7, -9, -27, -11, 4, 13, 14, 4, -5, -17, -53, -34, -21, -11, -28, -14, -24, -43 };
+const ENDGAME_KING_TABLE = [64]i32{ -74, -35, -18, -18, -11, 15, 4, -17, -12, 17, 14, 17, 17, 38, 23, 11, 10, 17, 23, 15, 20, 45, 44, 13, -8, 22, 24, 27, 26, 33, 26, 3, -18, -4, 21, 24, 27, 23, 9, -11, -19, -3, 11, 21, 23, 16, 7, -9, -27, -11, 4, 13, 14, 4, -5, -17, -53, -34, -21, -11, -28, -14, -24, -43 };
 
-const MIDGAME_VALUE = [6]i16{ 82, 337, 365, 477, 1025, 0 };
-const ENDGAME_VALUE = [6]i16{ 94, 281, 297, 512, 936, 0 };
+const MIDGAME_VALUE = [6]i32{ 82, 337, 365, 477, 1025, 0 };
+const ENDGAME_VALUE = [6]i32{ 94, 281, 297, 512, 936, 0 };
 
-const MIDGAME_TABLE = [6][64]i16{
+const MIDGAME_TABLE = [6][64]i32{
     MIDGAME_PAWN_TABLE,
     MIDGAME_KNIGHT_TABLE,
     MIDGAME_BISHOP_TABLE,
@@ -145,7 +145,7 @@ const MIDGAME_TABLE = [6][64]i16{
     MIDGAME_KING_TABLE,
 };
 
-const ENDGAME_TABLE = [6][64]i16{
+const ENDGAME_TABLE = [6][64]i32{
     ENDGAME_PAWN_TABLE,
     ENDGAME_KNIGHT_TABLE,
     ENDGAME_BISHOP_TABLE,
@@ -156,8 +156,8 @@ const ENDGAME_TABLE = [6][64]i16{
 
 const GAMEPHASE_INCREMENT = [6]u8{ 0, 1, 1, 2, 4, 0};
 
-var midgame_table: [2][6][64]i16 = undefined;
-var endgame_table: [2][6][64]i16 = undefined;
+var midgame_table: [2][6][64]i32 = undefined;
+var endgame_table: [2][6][64]i32 = undefined;
 
 pub fn init_tables() void {
     var piece: u5 = 0;
@@ -172,12 +172,12 @@ pub fn init_tables() void {
     }
 }
 
-pub fn evaluate(comptime active_color: Color, position: Position) i16 {
+pub fn evaluate(comptime active_color: Color, position: Position) i32 {
     const us = @enumToInt(active_color);
     const them = @enumToInt(active_color.other());
 
-    var midgame = [2]i16{ 0, 0 };
-    var endgame = [2]i16{ 0, 0 };
+    var midgame = [2]i32{ 0, 0 };
+    var endgame = [2]i32{ 0, 0 };
     var gamephase: u8 = 0;
 
     const pieces = position.as_array();
@@ -200,7 +200,6 @@ pub fn evaluate(comptime active_color: Color, position: Position) i16 {
 
     const midgamescore = midgame[us] - midgame[them];
     const endgamescore = endgame[us] - endgame[them];
-
     return @divTrunc(midgamescore * midgamephase + endgamescore * endgamephase, 24);
 }
 
@@ -208,5 +207,5 @@ test "starting position is equal" {
     const expectEqual = @import("std").testing.expectEqual;
 
     const position = Position.starting_position();
-    try expectEqual(@as(i16, 0), evaluate(Color.white, position));
+    try expectEqual(@as(i32, 0), evaluate(Color.white, position));
 }
